+++ lib/data/languages.dart
@@ -1,0 +1,542 @@
+import 'package:world_picker/model/model.dart';
+
+List<Language> supportedLanguages() {
+  return [
+    Language(code: 'aa', name: 'Afar', nativeName: 'Afaraf', countryCode: 'DJ'),
+    Language(
+        code: 'ab', name: 'Abkhazian', nativeName: 'Аҧсуа', countryCode: 'GE'),
+    Language(
+        code: 'ae', name: 'Avestan', nativeName: 'avesta', countryCode: 'IR'),
+    Language(
+        code: 'af',
+        name: 'Afrikaans',
+        nativeName: 'Afrikaans',
+        countryCode: 'ZA'),
+    Language(code: 'ak', name: 'Akan', nativeName: 'Akan', countryCode: 'GH'),
+    Language(
+        code: 'am', name: 'Amharic', nativeName: 'አማርኛ', countryCode: 'ET'),
+    Language(
+        code: 'an',
+        name: 'Aragonese',
+        nativeName: 'Aragonés',
+        countryCode: 'ES'),
+    Language(
+        code: 'ar', name: 'Arabic', nativeName: 'العربية', countryCode: 'SA'),
+    Language(
+        code: 'as', name: 'Assamese', nativeName: 'অসমীয়া', countryCode: 'IN'),
+    Language(
+        code: 'av', name: 'Avaric', nativeName: 'авар мацӀ', countryCode: 'RU'),
+    Language(
+        code: 'ay', name: 'Aymara', nativeName: 'aymar aru', countryCode: 'BO'),
+    Language(
+        code: 'az',
+        name: 'Azerbaijani',
+        nativeName: 'Azərbaycan dili',
+        countryCode: 'AZ'),
+    Language(
+        code: 'ba',
+        name: 'Bashkir',
+        nativeName: 'башҡорт теле',
+        countryCode: 'RU'),
+    Language(
+        code: 'be',
+        name: 'Belarusian',
+        nativeName: 'Беларуская',
+        countryCode: 'BY'),
+    Language(
+        code: 'bg',
+        name: 'Bulgarian',
+        nativeName: 'Български',
+        countryCode: 'BG'),
+    Language(
+        code: 'bh', name: 'Bihari', nativeName: 'भोजपुरी', countryCode: 'IN'),
+    Language(
+        code: 'bi', name: 'Bislama', nativeName: 'Bislama', countryCode: 'VU'),
+    Language(
+        code: 'bm',
+        name: 'Bambara',
+        nativeName: 'Bamanankan',
+        countryCode: 'ML'),
+    Language(
+        code: 'bn', name: 'Bengali', nativeName: 'বাংলা', countryCode: 'BD'),
+    Language(
+        code: 'bo', name: 'Tibetan', nativeName: 'བོད་ཡིག', countryCode: 'CN'),
+    Language(
+        code: 'br', name: 'Breton', nativeName: 'Brezhoneg', countryCode: 'FR'),
+    Language(
+        code: 'bs', name: 'Bosnian', nativeName: 'Bosanski', countryCode: 'BA'),
+    Language(
+        code: 'ca', name: 'Catalan', nativeName: 'Català', countryCode: 'ES'),
+    Language(
+        code: 'ce',
+        name: 'Chechen',
+        nativeName: 'Нохчийн мотт',
+        countryCode: 'RU'),
+    Language(
+        code: 'ch', name: 'Chamorro', nativeName: 'Chamoru', countryCode: 'GU'),
+    Language(
+        code: 'co', name: 'Corsican', nativeName: 'Corsu', countryCode: 'FR'),
+    Language(
+        code: 'cr', name: 'Cree', nativeName: 'ᓀᐦᐃᔭᐍᐏᐣ', countryCode: 'CA'),
+    Language(
+        code: 'cs', name: 'Czech', nativeName: 'Čeština', countryCode: 'CZ'),
+    Language(
+        code: 'cu',
+        name: 'Church Slavic',
+        nativeName: 'ѩзыкъ словѣньскъ',
+        countryCode: 'RU'),
+    Language(
+        code: 'cv',
+        name: 'Chuvash',
+        nativeName: 'чӑваш чӗлхи',
+        countryCode: 'RU'),
+    Language(
+        code: 'cy', name: 'Welsh', nativeName: 'Cymraeg', countryCode: 'GB'),
+    Language(
+        code: 'da', name: 'Danish', nativeName: 'Dansk', countryCode: 'DK'),
+    Language(
+        code: 'de', name: 'German', nativeName: 'Deutsch', countryCode: 'DE'),
+    Language(
+        code: 'dv', name: 'Divehi', nativeName: 'ދިވެހި', countryCode: 'MV'),
+    Language(
+        code: 'dz', name: 'Dzongkha', nativeName: 'རྫོང་ཁ', countryCode: 'BT'),
+    Language(code: 'ee', name: 'Ewe', nativeName: 'Eʋegbe', countryCode: 'GH'),
+    Language(
+        code: 'el', name: 'Greek', nativeName: 'Ελληνικά', countryCode: 'GR'),
+    Language(
+        code: 'en', name: 'English', nativeName: 'English', countryCode: 'US'),
+    Language(
+        code: 'eo',
+        name: 'Esperanto',
+        nativeName: 'Esperanto',
+        countryCode: null),
+    Language(
+        code: 'es', name: 'Spanish', nativeName: 'Español', countryCode: 'ES'),
+    Language(
+        code: 'et', name: 'Estonian', nativeName: 'Eesti', countryCode: 'EE'),
+    Language(
+        code: 'eu', name: 'Basque', nativeName: 'Euskara', countryCode: 'ES'),
+    Language(
+        code: 'fa', name: 'Persian', nativeName: 'فارسی', countryCode: 'IR'),
+    Language(
+        code: 'ff', name: 'Fulah', nativeName: 'Fulfulde', countryCode: 'SN'),
+    Language(
+        code: 'fi', name: 'Finnish', nativeName: 'Suomi', countryCode: 'FI'),
+    Language(
+        code: 'fj',
+        name: 'Fijian',
+        nativeName: 'Vosa Vakaviti',
+        countryCode: 'FJ'),
+    Language(
+        code: 'fo', name: 'Faroese', nativeName: 'Føroyskt', countryCode: 'FO'),
+    Language(
+        code: 'fr', name: 'French', nativeName: 'Français', countryCode: 'FR'),
+    Language(
+        code: 'fy',
+        name: 'Western Frisian',
+        nativeName: 'Frysk',
+        countryCode: 'NL'),
+    Language(
+        code: 'ga', name: 'Irish', nativeName: 'Gaeilge', countryCode: 'IE'),
+    Language(
+        code: 'gd',
+        name: 'Scottish Gaelic',
+        nativeName: 'Gàidhlig',
+        countryCode: 'GB'),
+    Language(
+        code: 'gl', name: 'Galician', nativeName: 'Galego', countryCode: 'ES'),
+    Language(
+        code: 'gn', name: 'Guarani', nativeName: 'Avañe\'ẽ', countryCode: 'PY'),
+    Language(
+        code: 'gu', name: 'Gujarati', nativeName: 'ગુજરાતી', countryCode: 'IN'),
+    Language(code: 'gv', name: 'Manx', nativeName: 'Gaelg', countryCode: 'IM'),
+    Language(code: 'ha', name: 'Hausa', nativeName: 'Hausa', countryCode: 'NG'),
+    Language(
+        code: 'he', name: 'Hebrew', nativeName: 'עברית', countryCode: 'IL'),
+    Language(
+        code: 'hi', name: 'Hindi', nativeName: 'हिन्दी', countryCode: 'IN'),
+    Language(
+        code: 'ho',
+        name: 'Hiri Motu',
+        nativeName: 'Hiri Motu',
+        countryCode: 'PG'),
+    Language(
+        code: 'hr',
+        name: 'Croatian',
+        nativeName: 'Hrvatski',
+        countryCode: 'HR'),
+    Language(
+        code: 'ht',
+        name: 'Haitian Creole',
+        nativeName: 'Kreyòl ayisyen',
+        countryCode: 'HT'),
+    Language(
+        code: 'hu', name: 'Hungarian', nativeName: 'Magyar', countryCode: 'HU'),
+    Language(
+        code: 'hy', name: 'Armenian', nativeName: 'Հայերեն', countryCode: 'AM'),
+    Language(
+        code: 'hz',
+        name: 'Herero',
+        nativeName: 'Otjiherero',
+        countryCode: 'NA'),
+    Language(
+        code: 'ia',
+        name: 'Interlingua',
+        nativeName: 'Interlingua',
+        countryCode: null),
+    Language(
+        code: 'id',
+        name: 'Indonesian',
+        nativeName: 'Bahasa Indonesia',
+        countryCode: 'ID'),
+    Language(
+        code: 'ie',
+        name: 'Interlingue',
+        nativeName: 'Occidental',
+        countryCode: null),
+    Language(code: 'ig', name: 'Igbo', nativeName: 'Igbo', countryCode: 'NG'),
+    Language(
+        code: 'ii',
+        name: 'Sichuan Yi',
+        nativeName: 'ꆈꌠ꒿ Nuosuhxop',
+        countryCode: 'CN'),
+    Language(
+        code: 'ik', name: 'Inupiaq', nativeName: 'Iñupiaq', countryCode: 'US'),
+    Language(code: 'io', name: 'Ido', nativeName: 'Ido', countryCode: null),
+    Language(
+        code: 'is',
+        name: 'Icelandic',
+        nativeName: 'Íslenska',
+        countryCode: 'IS'),
+    Language(
+        code: 'it', name: 'Italian', nativeName: 'Italiano', countryCode: 'IT'),
+    Language(
+        code: 'iu', name: 'Inuktitut', nativeName: 'ᐃᓄᒃᑎᑐᑦ', countryCode: 'CA'),
+    Language(
+        code: 'ja', name: 'Japanese', nativeName: '日本語', countryCode: 'JP'),
+    Language(
+        code: 'jv',
+        name: 'Javanese',
+        nativeName: 'Basa Jawa',
+        countryCode: 'ID'),
+    Language(
+        code: 'ka', name: 'Georgian', nativeName: 'ქართული', countryCode: 'GE'),
+    Language(
+        code: 'kg', name: 'Kongo', nativeName: 'Kikongo', countryCode: 'CG'),
+    Language(
+        code: 'ki', name: 'Kikuyu', nativeName: 'Gĩkũyũ', countryCode: 'KE'),
+    Language(
+        code: 'kj',
+        name: 'Kwanyama',
+        nativeName: 'Kuanyama',
+        countryCode: 'AO'),
+    Language(
+        code: 'kk',
+        name: 'Kazakh',
+        nativeName: 'Қазақ тілі',
+        countryCode: 'KZ'),
+    Language(
+        code: 'kl',
+        name: 'Kalaallisut',
+        nativeName: 'Kalaallisut',
+        countryCode: 'GL'),
+    Language(
+        code: 'km', name: 'Khmer', nativeName: 'ភាសាខ្មែរ', countryCode: 'KH'),
+    Language(
+        code: 'kn', name: 'Kannada', nativeName: 'ಕನ್ನಡ', countryCode: 'IN'),
+    Language(code: 'ko', name: 'Korean', nativeName: '한국어', countryCode: 'KR'),
+    Language(
+        code: 'kr', name: 'Kanuri', nativeName: 'Kanuri', countryCode: 'NG'),
+    Language(
+        code: 'ks',
+        name: 'Kashmiri',
+        nativeName: 'कश्मीरी / كشميري',
+        countryCode: 'IN'),
+    Language(
+        code: 'ku',
+        name: 'Kurdish',
+        nativeName: 'Kurdî / كوردی',
+        countryCode: 'IQ'),
+    Language(code: 'kv', name: 'Komi', nativeName: 'Коми', countryCode: 'RU'),
+    Language(
+        code: 'kw', name: 'Cornish', nativeName: 'Kernewek', countryCode: 'GB'),
+    Language(
+        code: 'ky', name: 'Kirghiz', nativeName: 'Кыргызча', countryCode: 'KG'),
+    Language(
+        code: 'la', name: 'Latin', nativeName: 'Latine', countryCode: null),
+    Language(
+        code: 'lb',
+        name: 'Luxembourgish',
+        nativeName: 'Lëtzebuergesch',
+        countryCode: 'LU'),
+    Language(
+        code: 'lg', name: 'Ganda', nativeName: 'Luganda', countryCode: 'UG'),
+    Language(
+        code: 'li',
+        name: 'Limburgish',
+        nativeName: 'Limburgs',
+        countryCode: 'NL'),
+    Language(
+        code: 'ln', name: 'Lingala', nativeName: 'Lingála', countryCode: 'CD'),
+    Language(code: 'lo', name: 'Lao', nativeName: 'ພາສາລາວ', countryCode: 'LA'),
+    Language(
+        code: 'lt',
+        name: 'Lithuanian',
+        nativeName: 'Lietuvių',
+        countryCode: 'LT'),
+    Language(
+        code: 'lu',
+        name: 'Luba-Katanga',
+        nativeName: 'Tshiluba',
+        countryCode: 'CD'),
+    Language(
+        code: 'lv', name: 'Latvian', nativeName: 'Latviešu', countryCode: 'LV'),
+    Language(
+        code: 'mg',
+        name: 'Malagasy',
+        nativeName: 'Malagasy',
+        countryCode: 'MG'),
+    Language(
+        code: 'mh',
+        name: 'Marshallese',
+        nativeName: 'Kajin M̧ajeļ',
+        countryCode: 'MH'),
+    Language(code: 'mi', name: 'Maori', nativeName: 'Māori', countryCode: 'NZ'),
+    Language(
+        code: 'mk',
+        name: 'Macedonian',
+        nativeName: 'Македонски',
+        countryCode: 'MK'),
+    Language(
+        code: 'ml', name: 'Malayalam', nativeName: 'മലയാളം', countryCode: 'IN'),
+    Language(
+        code: 'mn', name: 'Mongolian', nativeName: 'Монгол', countryCode: 'MN'),
+    Language(
+        code: 'mr', name: 'Marathi', nativeName: 'मराठी', countryCode: 'IN'),
+    Language(
+        code: 'ms',
+        name: 'Malay',
+        nativeName: 'Bahasa Melayu',
+        countryCode: 'MY'),
+    Language(
+        code: 'mt', name: 'Maltese', nativeName: 'Malti', countryCode: 'MT'),
+    Language(
+        code: 'my', name: 'Burmese', nativeName: 'မြန်မာစာ', countryCode: 'MM'),
+    Language(
+        code: 'na',
+        name: 'Nauru',
+        nativeName: 'Ekakairũ Naoero',
+        countryCode: 'NR'),
+    Language(
+        code: 'nb',
+        name: 'Norwegian Bokmål',
+        nativeName: 'Norsk Bokmål',
+        countryCode: 'NO'),
+    Language(
+        code: 'nd',
+        name: 'North Ndebele',
+        nativeName: 'Sindebele',
+        countryCode: 'ZW'),
+    Language(
+        code: 'ne', name: 'Nepali', nativeName: 'नेपाली', countryCode: 'NP'),
+    Language(
+        code: 'ng', name: 'Ndonga', nativeName: 'Owambo', countryCode: 'NA'),
+    Language(
+        code: 'nl', name: 'Dutch', nativeName: 'Nederlands', countryCode: 'NL'),
+    Language(
+        code: 'nn',
+        name: 'Norwegian Nynorsk',
+        nativeName: 'Norsk Nynorsk',
+        countryCode: 'NO'),
+    Language(
+        code: 'no', name: 'Norwegian', nativeName: 'Norsk', countryCode: 'NO'),
+    Language(
+        code: 'nr',
+        name: 'South Ndebele',
+        nativeName: 'Ndebele',
+        countryCode: 'ZA'),
+    Language(
+        code: 'nv',
+        name: 'Navajo',
+        nativeName: 'Diné Bizaad',
+        countryCode: 'US'),
+    Language(
+        code: 'ny',
+        name: 'Chichewa',
+        nativeName: 'ChiCheŵa',
+        countryCode: 'MW'),
+    Language(
+        code: 'oc', name: 'Occitan', nativeName: 'Occitan', countryCode: 'FR'),
+    Language(
+        code: 'oj', name: 'Ojibwa', nativeName: 'ᐊᓂᔑᓈᐯᒧᐎᓐ', countryCode: 'CA'),
+    Language(
+        code: 'om',
+        name: 'Oromo',
+        nativeName: 'Afaan Oromoo',
+        countryCode: 'ET'),
+    Language(code: 'or', name: 'Odia', nativeName: 'ଓଡ଼ିଆ', countryCode: 'IN'),
+    Language(
+        code: 'os', name: 'Ossetian', nativeName: 'Ирон', countryCode: 'GE'),
+    Language(
+        code: 'pa', name: 'Punjabi', nativeName: 'ਪੰਜਾਬੀ', countryCode: 'IN'),
+    Language(code: 'pi', name: 'Pali', nativeName: 'पालि', countryCode: null),
+    Language(
+        code: 'pl', name: 'Polish', nativeName: 'Polski', countryCode: 'PL'),
+    Language(code: 'ps', name: 'Pashto', nativeName: 'پښتو', countryCode: 'AF'),
+    Language(
+        code: 'pt',
+        name: 'Portuguese',
+        nativeName: 'Português',
+        countryCode: 'PT'),
+    Language(
+        code: 'pt-BR',
+        name: 'Portuguese (Brazil)',
+        nativeName: 'Português (Brasil)',
+        countryCode: 'BR'),
+    Language(
+        code: 'qu',
+        name: 'Quechua',
+        nativeName: 'Runa Simi',
+        countryCode: 'PE'),
+    Language(
+        code: 'rm',
+        name: 'Romansh',
+        nativeName: 'Rumantsch',
+        countryCode: 'CH'),
+    Language(
+        code: 'rn', name: 'Rundi', nativeName: 'Ikirundi', countryCode: 'BI'),
+    Language(
+        code: 'ro', name: 'Romanian', nativeName: 'Română', countryCode: 'RO'),
+    Language(
+        code: 'ru', name: 'Russian', nativeName: 'Русский', countryCode: 'RU'),
+    Language(
+        code: 'rw',
+        name: 'Kinyarwanda',
+        nativeName: 'Ikinyarwanda',
+        countryCode: 'RW'),
+    Language(
+        code: 'sa',
+        name: 'Sanskrit',
+        nativeName: 'संस्कृतम्',
+        countryCode: 'IN'),
+    Language(
+        code: 'sc', name: 'Sardinian', nativeName: 'Sardu', countryCode: 'IT'),
+    Language(
+        code: 'sd',
+        name: 'Sindhi',
+        nativeName: 'सिन्धी / سنڌي',
+        countryCode: 'PK'),
+    Language(
+        code: 'se',
+        name: 'Northern Sami',
+        nativeName: 'Davvisámegiella',
+        countryCode: 'NO'),
+    Language(code: 'sg', name: 'Sango', nativeName: 'Sängö', countryCode: 'CF'),
+    Language(
+        code: 'si', name: 'Sinhala', nativeName: 'සිංහල', countryCode: 'LK'),
+    Language(
+        code: 'sk',
+        name: 'Slovak',
+        nativeName: 'Slovenčina',
+        countryCode: 'SK'),
+    Language(
+        code: 'sl',
+        name: 'Slovenian',
+        nativeName: 'Slovenščina',
+        countryCode: 'SI'),
+    Language(
+        code: 'sm',
+        name: 'Samoan',
+        nativeName: 'Gagana Samoa',
+        countryCode: 'WS'),
+    Language(
+        code: 'sn', name: 'Shona', nativeName: 'chiShona', countryCode: 'ZW'),
+    Language(
+        code: 'so', name: 'Somali', nativeName: 'Soomaali', countryCode: 'SO'),
+    Language(
+        code: 'sq', name: 'Albanian', nativeName: 'Shqip', countryCode: 'AL'),
+    Language(
+        code: 'sr', name: 'Serbian', nativeName: 'Српски', countryCode: 'RS'),
+    Language(
+        code: 'ss', name: 'Swati', nativeName: 'SiSwati', countryCode: 'SZ'),
+    Language(
+        code: 'st',
+        name: 'Southern Sotho',
+        nativeName: 'Sesotho',
+        countryCode: 'LS'),
+    Language(
+        code: 'su',
+        name: 'Sundanese',
+        nativeName: 'Basa Sunda',
+        countryCode: 'ID'),
+    Language(
+        code: 'sv', name: 'Swedish', nativeName: 'Svenska', countryCode: 'SE'),
+    Language(
+        code: 'sw',
+        name: 'Swahili',
+        nativeName: 'Kiswahili',
+        countryCode: 'KE'),
+    Language(code: 'ta', name: 'Tamil', nativeName: 'தமிழ்', countryCode: 'IN'),
+    Language(
+        code: 'te', name: 'Telugu', nativeName: 'తెలుగు', countryCode: 'IN'),
+    Language(
+        code: 'tg', name: 'Tajik', nativeName: 'тоҷикӣ', countryCode: 'TJ'),
+    Language(code: 'th', name: 'Thai', nativeName: 'ไทย', countryCode: 'TH'),
+    Language(
+        code: 'ti', name: 'Tigrinya', nativeName: 'ትግርኛ', countryCode: 'ER'),
+    Language(
+        code: 'tk', name: 'Turkmen', nativeName: 'Türkmen', countryCode: 'TM'),
+    Language(
+        code: 'tl', name: 'Tagalog', nativeName: 'Tagalog', countryCode: 'PH'),
+    Language(
+        code: 'tn', name: 'Tswana', nativeName: 'Setswana', countryCode: 'BW'),
+    Language(
+        code: 'to', name: 'Tonga', nativeName: 'Faka Tonga', countryCode: 'TO'),
+    Language(
+        code: 'tr', name: 'Turkish', nativeName: 'Türkçe', countryCode: 'TR'),
+    Language(
+        code: 'ts', name: 'Tsonga', nativeName: 'Xitsonga', countryCode: 'ZA'),
+    Language(
+        code: 'tt', name: 'Tatar', nativeName: 'татарча', countryCode: 'RU'),
+    Language(code: 'tw', name: 'Twi', nativeName: 'Twi', countryCode: 'GH'),
+    Language(
+        code: 'ty',
+        name: 'Tahitian',
+        nativeName: 'Reo Tahiti',
+        countryCode: 'PF'),
+    Language(
+        code: 'ug', name: 'Uighur', nativeName: 'ئۇيغۇرچە', countryCode: 'CN'),
+    Language(
+        code: 'uk',
+        name: 'Ukrainian',
+        nativeName: 'Українська',
+        countryCode: 'UA'),
+    Language(code: 'ur', name: 'Urdu', nativeName: 'اردو', countryCode: 'PK'),
+    Language(
+        code: 'uz', name: 'Uzbek', nativeName: 'Oʻzbekcha', countryCode: 'UZ'),
+    Language(
+        code: 've', name: 'Venda', nativeName: 'Tshivenḓa', countryCode: 'ZA'),
+    Language(
+        code: 'vi',
+        name: 'Vietnamese',
+        nativeName: 'Tiếng Việt',
+        countryCode: 'VN'),
+    Language(
+        code: 'vo', name: 'Volapük', nativeName: 'Volapük', countryCode: null),
+    Language(
+        code: 'wa', name: 'Walloon', nativeName: 'Walon', countryCode: 'BE'),
+    Language(
+        code: 'wo', name: 'Wolof', nativeName: 'Wollof', countryCode: 'SN'),
+    Language(
+        code: 'xh', name: 'Xhosa', nativeName: 'isiXhosa', countryCode: 'ZA'),
+    Language(
+        code: 'yi', name: 'Yiddish', nativeName: 'ייִדיש', countryCode: 'DE'),
+    Language(
+        code: 'yo', name: 'Yoruba', nativeName: 'Yorùbá', countryCode: 'NG'),
+    Language(
+        code: 'za', name: 'Zhuang', nativeName: 'Cuengh', countryCode: 'CN'),
+    Language(code: 'zh', name: 'Chinese', nativeName: '中文', countryCode: 'CN'),
+    Language(
+        code: 'zu', name: 'Zulu', nativeName: 'isiZulu', countryCode: 'ZA'),
+  ];
+}